Move getters of cxprogram to its file
